organized actions to its own package in controllers and fixed bug where clicking on the same note would generate actions
