+++ src/proj5ZhouRinkerSahChistolini/Controllers/Actions/DeleteNoteAction.java
@@ -1,8 +1,8 @@
-package proj5ZhouRinkerSahChistolini.Controllers;
+package proj5ZhouRinkerSahChistolini.Controllers.Actions;
 
 import javafx.scene.Node;
+import proj5ZhouRinkerSahChistolini.Controllers.CompositionPanelController;
 import proj5ZhouRinkerSahChistolini.Models.Note;
-import proj5ZhouRinkerSahChistolini.Views.NoteRectangle;
 import proj5ZhouRinkerSahChistolini.Views.SelectableRectangle;
 
 import java.util.Collection;
