+++ src/proj5ZhouRinkerSahChistolini/Controllers/Actions/ExtendNoteAction.java
@@ -1,7 +1,5 @@
-package proj5ZhouRinkerSahChistolini.Controllers;
+package proj5ZhouRinkerSahChistolini.Controllers.Actions;
 
-import javafx.scene.Node;
-import proj5ZhouRinkerSahChistolini.Models.Note;
 import proj5ZhouRinkerSahChistolini.Views.SelectableRectangle;
 
 import java.util.Collection;
