+++ src/proj5ZhouRinkerSahChistolini/Controllers/CompositionPanelController.java
@@ -21,6 +21,9 @@ import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Pane;
 import javafx.scene.shape.Line;
 import javafx.scene.shape.Rectangle;
+import proj5ZhouRinkerSahChistolini.Controllers.Actions.Actionable;
+import proj5ZhouRinkerSahChistolini.Controllers.Actions.GroupNoteAction;
+import proj5ZhouRinkerSahChistolini.Controllers.Actions.UngroupNoteAction;
 import proj5ZhouRinkerSahChistolini.Models.Note;
 import proj5ZhouRinkerSahChistolini.Views.GroupRectangle;
 import proj5ZhouRinkerSahChistolini.Models.Composition;
